refactor(data-wallpaper): lift compose/wallpaper payload schema to data-wallpaper-core
Extract WallpaperPayload, KIND, and SCHEMA_VERSION out of
WallpaperDataProduct.kt into a new data-wallpaper-core module so MCP
clients and other connectors can depend on the schema without pulling
in the daemon-side WallpaperDataProductRegistry, Compose Material3,
or material-kolor.

WallpaperDataProductRegistry.KIND / .SCHEMA_VERSION stay as forwarding
constants pointing at Material3WallpaperProduct so existing in-tree
call sites (WallpaperOverrideExtension.ID, tests) keep working.

data-wallpaper-core api-depends on :daemon:core because WallpaperPayload
references the protocol-level WallpaperPaletteStyle enum. MCP clients
already consume the protocol module, so this just keeps the wallpaper
schema reachable alongside it.

Mirrors the core/connector split that fonts, strings, layoutinspector,
resources, a11y, history, and theme already use. No behavior change.
